Move public headers to include directory
