+++ include/meson.build
@@ -1,0 +1,1 @@
+install_headers('evavm/disassemble.h', 'evavm/eva.h', subdir:'evavm')
