+++ src/disassemble.c
@@ -1,5 +1,5 @@
-#include "disassemble.h"
-#include "eva.h"
+#include <evavm/disassemble.h>
+#include <evavm/eva.h>
 
 void disassemble(opcode_t *op) { disassemble_file(stdout, op); }
